Corrected the code and added answers where required.
For the final plot:I did not remove grid lines as the figure becomes harder to read especially with empty lines for countries not in that facet. I left both facets and color legend as requested, though I think it is redundant. If anything, I think the intial version without facets works best as it summarised everything in one plot.